pulled database pool creation into its own file so it can be used during uploadRoutes as well
